feat(cobolnet): N1 — the Int128 carrier: expression pipeline + runtime engine Int128-monomorphic (numeric design D1+D2 shipped)
Wave 1 of per-edition numerics end-to-end (owner directive). NO numeric
intermediate computes in long anywhere:
- Runtime: every CobolNum kernel Int128-typed (Rescale/Store/TryStore/
  MulChecked=38-digit-escape/Divide/DivideOrThrow/Format*/ParseDisplay/
  FromAlphanumeric); CobolEdit.Format takes Int128; storage stays narrow
  native long (wide tier = N2).
- Renderer: Combine forces ((Int128)(a) op (b)) on every operation;
  Real() casts explicitly for the float path. D2 SHIPPED: nested
  divisions carry DivGuardDigits=14 (clamped to the 38-digit alignment
  headroom); outermost keeps exact single-rounding-at-receiver.
- Emitter: ONE width-aware Narrow() at the store boundary; long temps
  cast at their boundaries; DIVIDE-REMAINDER subsidiary quotient is
  Int128; generated programs gain 'using System;'.
- Doc-sync: D1/D2 SHIPPED notes incl. the deliberate refinement —
  CobolInt.Scale resolved at COMPILE time (static scales + monomorphic
  kernels), rationale recorded.
- DEVLOG 539.

Byte-identical under the new engine: 440 conformance (all 54 NC) + 15
unit green. COMPUTE c = a*b on two PIC 9(18) operands is now exact.

+++ src/Cobol.Net.Runtime/Tables/CobolTable.cs
@@ -38,5 +38,5 @@ public static class CobolTable
     public static long Occ(long value) => value;
 
     /// <inheritdoc cref="Occ(long)"/>
-    public static long Occ(string image) => CobolNum.FromAlphanumeric(image);
+    public static long Occ(string image) => (long)CobolNum.FromAlphanumeric(image);
 }
